+++ tests/__init__.py
@@ -1,0 +1,25 @@
+# -*- coding: utf-8 -*-
+"""
+tkcalendar - Calendar and DateEntry widgets for Tkinter
+Copyright 2017-2018 Juliette Monsel <j_4321@protonmail.com>
+
+tkcalendar is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+tkcalendar is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+Test
+"""
+
+from tests.base_widget_testcase import BaseWidgetTest, TestEvent, tk, ttk
+import locale
+
+locale.setlocale(locale.LC_ALL, '')
