+++ tests/test_calendar.py
@@ -1,60 +1,6 @@
-# -*- coding: utf-8 -*-
-"""
-tkcalendar - Calendar and DateEntry widgets for Tkinter
-Copyright 2017-2018 Juliette Monsel <j_4321@protonmail.com>
-
-tkcalendar is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-tkcalendar is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-Test
-"""
-
-
-import unittest
-from tkcalendar import Calendar, DateEntry
+from tests import BaseWidgetTest, TestEvent, tk, ttk
+from tkcalendar import Calendar
 from datetime import datetime, date
-import locale
-try:
-    import Tkinter as tk
-    import ttk
-except ImportError:
-    import tkinter as tk
-    from tkinter import ttk
-from pynput.mouse import Controller, Button
-
-locale.setlocale(locale.LC_ALL, '')
-
-
-class BaseWidgetTest(unittest.TestCase):
-    def setUp(self):
-        self.window = tk.Toplevel()
-        self.window.update()
-
-    def tearDown(self):
-        self.window.update()
-        self.window.destroy()
-
-
-class TestEvent:
-    """Fake event for testing."""
-    def __init__(self, **kwargs):
-        self._prop = kwargs
-
-    def __getattr__(self, attr):
-        if attr not in self._prop:
-            raise AttributeError("TestEvent has no attribute %s." % attr)
-        else:
-            return self._prop[attr]
 
 
 class TestCalendar(BaseWidgetTest):
@@ -259,111 +205,3 @@ class TestCalendar(BaseWidgetTest):
         self.window.update()
         for op in options[7:]:
             self.assertEqual(widget.cget(op), "yellow")
-
-
-class TestDateEntry(BaseWidgetTest):
-    def test_dateentry_init(self):
-        widget = DateEntry(self.window, width=12, background='darkblue',
-                           foreground='white', borderwidth=2)
-        widget.pack()
-        self.window.update()
-        widget.destroy()
-        widget = DateEntry(self.window, year=2012)
-        widget.pack()
-        self.window.update()
-        widget.destroy()
-
-    def test_dateentry_drop_down(self):
-        """Check whether drop down opens on click."""
-        widget = DateEntry(self.window)
-        widget.pack()
-        self.window.update()
-        x, y = widget.winfo_rootx(), widget.winfo_rooty()
-        w = widget.winfo_width()
-        mouse_controller = Controller()
-        mouse_controller.position = (x + w - 2, y + 2)
-        self.window.update()
-        mouse_controller.press(Button.left)
-        self.window.update()
-        mouse_controller.release(Button.left)
-        self.window.update()
-        self.assertTrue(widget._top_cal.winfo_ismapped())
-
-    def test_dateentry_get_set(self):
-        widget = DateEntry(self.window, width=12, background='darkblue',
-                           foreground='white', borderwidth=2, font='Arial 9')
-        widget.pack()
-        self.window.update()
-
-        keys = ['exportselection',
-                'invalidcommand',
-                'justify',
-                'show',
-                'cursor',
-                'style',
-                'state',
-                'takefocus',
-                'textvariable',
-                'validate',
-                'validatecommand',
-                'width',
-                'xscrollcommand']
-        keys.extend(widget._calendar.keys())
-        self.assertEqual(sorted(list(set(keys))), sorted(widget.keys()))
-
-        self.assertEqual(widget["background"], 'darkblue')
-        self.assertEqual(widget.cget("width"), 12)
-
-        widget["borderwidth"] = 5
-        self.window.update()
-        self.assertEqual(widget["borderwidth"], 5)
-
-        widget.config(font="Arial 20 bold")
-        self.window.update()
-        self.assertEqual(widget["font"], "Arial 20 bold")
-
-        widget.config(style="my.TEntry")
-        self.window.update()
-        self.assertEqual(widget["style"], "my.TEntry")
-
-    def test_dateentry_functions(self):
-        widget = DateEntry(self.window, width=12, background='darkblue',
-                           foreground='white', borderwidth=2)
-        widget.pack()
-        self.window.update()
-
-        widget.set_date(datetime(2018, 12, 31).strftime('%x'))
-        self.assertEqual(widget.get_date(), date(2018, 12, 31))
-        with self.assertRaises(ValueError):
-            widget.set_date("ab")
-        widget.set_date(datetime(2015, 12, 31))
-        self.assertEqual(widget.get_date(), date(2015, 12, 31))
-        self.assertEqual(widget.get(), datetime(2015, 12, 31).strftime("%x"))
-
-        widget.delete(0, "end")
-        widget.insert(0, "abc")
-        self.window.focus_force()
-        self.assertEqual(widget.get_date(), date(2015, 12, 31))
-
-        widget._on_motion(TestEvent(x=10, y=20))
-        widget._on_b1_press(TestEvent(x=10, y=20))
-        widget._on_b1_press(TestEvent(x=widget.winfo_width() - 2, y=2))
-        widget._on_focus_out_cal(TestEvent(x=10, y=20))
-
-        widget.state(("disabled",))
-        self.window.update()
-        self.assertIn("disabled", widget.state())
-
-        widget.drop_down()
-        self.window.update()
-        widget._select()
-        self.window.update()
-        widget.drop_down()
-        self.window.update()
-        widget.drop_down()
-        self.window.update()
-
-        widget.configure(state='readonly')
-        self.window.update()
-        widget._select()
-        self.assertIn('readonly', widget.state())
